Age-out tiers, recovery-after-outage pattern, ESP32 big profile, docs
- PROTOCOL.md: §10 age-out tiers now 72h/4h/5min (was 48h/1h/5min),
  §4.3 clarifies self-query (target == own sender_id) is allowed,
  §17 rewrites versioning policy (no version field, ever — new port
  for new protocol), Appendix A added: recovery-after-total-outage
  pattern using the 128-byte node payload and advertise-verify-act
  invariant (DRBD UUID swap example included).

- python/echo/witness.py + crates/bedrock-echo-witness/src/state.rs:
  age-out tier constants match spec; Rust widens NodeEntry.cluster_slot
  u8 → u16 so bigger impls can address >256 clusters.

- firmware/esp32-c: profile picked for Olimex ESP32-POE-ISO at
  256 nodes × 128 clusters × 192 rate entries (~60 KB .bss, ~107 KB
  heap headroom). Previous 512×256×256 attempt left only ~50 KB heap
  and also hit a stack-overflow bug in the STATUS_LIST handler
  (entries[ECHO_MAX_NODES] on an 8 KB task stack). Fixed the stack
  array to ECHO_LIST_MAX_ENTRIES (64) which matches what the loop
  actually fills.

- docs/bedrock-cluster-log.md: draft design for the optional
  hash-chained cluster-transition log (separate concern from Echo
  itself, lives on DRBD nodes not the witness).

- docs/adoption/README.md: per-system notes covering DRBD, Corosync
  qdevice, Raft family (etcd/RQlite/Consul/CockroachDB), Galera,
  Patroni, Redis Sentinel, ZooKeeper, Ceph, MinIO, Nutanix Metro,
  VMware vSAN — where Echo fits, where it doesn't, what integration
  looks like.

+++ crates/bedrock-echo-witness/src/handler.rs
@@ -73,7 +73,7 @@ fn handle_bootstrap(state: &mut State, data: &[u8], src_ipv4: [u8; 4], now_ms: u
                 in_use: true,
                 sender_id: d.header.sender_id,
                 sender_ipv4: src_ipv4,
-                cluster_slot: cs as u8,
+                cluster_slot: cs as u16,
                 last_rx_ms: now_ms,
                 last_rx_sequence: 0,
                 last_tx_sequence: 0,
